# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Backend API",
        "description": "BillPays API RESTful",
        "contact": {
            "name": "Vitor Rodriuues",
            "url": "https://github.com/vs0uz4/cr_laravel_com_android",
            "email": "vitor.rodrigues@gmail.com"
        },
        "version": "0.0.1"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request an JWT Token",
                "description": "Request an JWT Token",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created JWT Token"
                    },
                    "401": {
                        "description": "Invalid Credentials"
                    },
                    "500": {
                        "description": "Could not Create Token"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Revoke a JWT Token",
                "description": "Revoke a JWT Token",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request - Invalid Token"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid Token"
                    },
                    "500": {
                        "description": "Could not Invalidate Token"
                    }
                }
            }
        },
        "/refresh_token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh a JWT Token",
                "description": "Refresh a JWT Token",
                "operationId": "refresh_token",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a new JWT Token"
                    },
                    "500": {
                        "description": "Could not Refresh Token"
                    }
                }
            }
        },
...
    },
...
}